# openapi/v1/schema.yaml
openapi: 3.1.0
info:
  title: Client Dashboard API
  version: 1.0.0
  description: >
    Backend API for Client Dashboard.
    Provides authentication and user management for web and mobile clients.

servers:
  - url: http://localhost:4000
    description: Local development

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Profile
    description: Profile endpoints
  - name: Orders
    description: Orders endpoints

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisteredResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/requestcode:
    get:
      tags: [Auth]
      summary: Request account activation code
      security: []
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
            example: a@a.com
          description: User email to get a code
      responses:
        '200':
          description: Successfully sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivationCodeSentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/activate:
    get:
      tags: [Auth]
      summary: Activate account with provided code
      security: []
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
            example: 086a0
          description: Activation code
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
            example: user@email.com
      responses:
        '200':
          description: Account activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountActivatedResponse'
        '400':
          description: Activation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=abc123; HttpOnly; Secure; SameSite=Strict
              description: HTTP-only cookie containing the refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Non-activated account log-in attempt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using refresh cookie
      security: []
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/resetpassword:
    get:
      tags: [Auth]
      summary: Reset password
      security: []
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
            example: user@email.com
      responses:
        '204':
          description: Reset password link sent
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/verifyresetcode:
    get:
      tags: [Auth]
      description: Verify reset code
      security: []
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
            example: user@email.com
        - in: query
          name: code
          required: true
          schema:
            type: string
            example: AJH412
      responses:
        '204':
          description: Reset successful
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Wrong code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout and clear refresh token cookie
      responses:
        '204':
          description: Successfully logged out
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile/me:
    get:
      tags: [Profile]
      summary: Request profile data
      responses:
        '200':
          description: Profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountInfoResponse'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile/changepassword:
    post:
      tags: [Profile]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed successfully
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden, oldPassword is not valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders:
    get:
      tags: [Orders]
      summary: Request orders
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden. Blocked, inactive or unverified user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  parameters:
    PageParam:
      in: query
      name: page
      required: true
      schema:
        type: integer
        minimum: 1
        example: 1
      description: Page
    PageSizeParam:
      in: query
      name: pageSize
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 50
        example: 10
      description: Page size
    StatusParam:
      in: query
      name: status
      required: false
      schema:
        $ref: '#/components/schemas/OrderStatus'
      description: Order status filter
    SortParam:
      in: query
      name: sort
      required: false
      schema:
        $ref: '#/components/schemas/OrderSort'
      description: Sort order
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          example: MySecurePass123
        name:
          type: string
          example: John Doe

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          example: MySecurePass123

    ChangePasswordRequest:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          type: string
          example: MySecurePass123
        newPassword:
          type: string
          minLength: 8
          example: MyNewSecurePass123

    RegisteredResponse:
      type: object
      properties:
        message:
          type: string
          example: User registered

    AuthResponse:
      type: object
      required: [user, accessToken]
      properties:
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          type: string
          description: Short-lived JWT access token

    AccountInfoResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'

    ActivationCodeSentResponse:
      type: object
      properties:
        message:
          type: string
          example: Activation code sent

    AccountActivatedResponse:
      type: object
      properties:
        message:
          type: string
          example: Account activated

    AccessTokenResponse:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
          description: New JWT access token

    User:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        verifiedClient:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - email
        - verifiedClient
        - createdAt

    OrderStatus:
      type: string
      enum: [pending, processing, shipped, delivered, canceled]

    OrderSort:
      type: string
      enum: [dateAsc, dateDesc, priceAsc, priceDesc]

    Order:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        date:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/OrderStatus'
        title:
          type: string
      required: [id, date, status, title, code]

    OrdersResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        totalPages:
          type: integer
      required: [orders, totalPages]

    ErrorResponse:
      type: object
      required: [statusCode, error, message]
      properties:
        statusCode:
          type: integer
          example: 401
        error:
          type: string
          example: Unauthorized
        message:
          type: string
          example: Invalid email or password
        details:
          type: object
          additionalProperties: true
          nullable: true
          description: Optional additional error context (validation errors, etc.)

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []